Moving sum_operation to arithmetics plugin
