+++ phylanx/plugins/matrixops/matrixops.hpp
@@ -32,7 +32,6 @@
 #include <phylanx/plugins/matrixops/shuffle_operation.hpp>
 #include <phylanx/plugins/matrixops/slicing_operation.hpp>
 #include <phylanx/plugins/matrixops/squeeze_operation.hpp>
-#include <phylanx/plugins/matrixops/sum_operation.hpp>
 #include <phylanx/plugins/matrixops/transpose_operation.hpp>
 #include <phylanx/plugins/matrixops/unique.hpp>
 #include <phylanx/plugins/matrixops/vstack_operation.hpp>
