+++ tests/unit/plugins/arithmetics/CMakeLists.txt
@@ -10,6 +10,7 @@ set(tests
     mul_operation
     prod_operation
     sub_operation
+    sum_operation
     unary_minus_operation
    )
