+++ tests/unit/plugins/arithmetics/sum_operation.cpp
